Use flask like its actually flask
