+++ main.py
@@ -1,4 +1,9 @@
-from blaise_nisra_case_mover import app
+import os
+
+from app.blaise_nisra_case_mover import app, load_config
+
+load_config(app)
 
 if __name__ == "__main__":
-    app.run(host="0.0.0.0")
+    port = os.getenv("5000")
+    app.run(host="0.0.0.0", port=port)
